refactor: Move draft-related utilities to a new `drafts.ts` file
