+++ client/src/utils/index.ts
@@ -6,6 +6,7 @@ export * from './files';
 export * from './latex';
 export * from './theme';
 export * from './forms';
+export * from './drafts';
 export * from './convos';
 export * from './presets';
 export * from './prompts';
